refactor: oneToOne is optional 엄격한 1대1은 선택적

+++ packages/socket-server/src/utils/bi-map.class.ts
@@ -1,18 +1,22 @@
 export class BiMap<L = any, R = any> {
   private leftToRight: Map<L, R>;
   private rightToLeft: Map<R, L>;
-  constructor() {
+  private isOneToOne: boolean;
+  constructor(opts?: { isOneToOne?: boolean }) {
     this.leftToRight = new Map<L, R>();
     this.rightToLeft = new Map<R, L>();
+    this.isOneToOne = !!opts?.isOneToOne;
   }
   set(left: L, right: R): void {
-    const prevRight = this.getByLeft(left);
-    const prevLeft = this.getByRight(right);
-    if (prevRight !== undefined) {
-      this.deleteByRight(prevRight);
-    }
-    if (prevLeft !== undefined) {
-      this.deleteByLeft(prevLeft);
+    if (this.isOneToOne) {
+      const prevRight = this.getByLeft(left);
+      const prevLeft = this.getByRight(right);
+      if (prevRight !== undefined) {
+        this.deleteByRight(prevRight);
+      }
+      if (prevLeft !== undefined) {
+        this.deleteByLeft(prevLeft);
+      }
     }
     this.leftToRight.set(left, right);
     this.rightToLeft.set(right, left);
@@ -31,13 +35,17 @@ export class BiMap<L = any, R = any> {
   }
   deleteByLeft(left: L): void {
     const right = this.getByLeft(left);
-    this.leftToRight.delete(left);
-    this.rightToLeft.delete(right);
+    if (this.isOneToOne) {
+      this.leftToRight.delete(left);
+      this.rightToLeft.delete(right);
+    }
   }
   deleteByRight(right: R): void {
     const left = this.getByRight(right);
-    this.rightToLeft.delete(right);
-    this.leftToRight.delete(left);
+    if (this.isOneToOne) {
+      this.rightToLeft.delete(right);
+      this.leftToRight.delete(left);
+    }
   }
   hasByLeft(left: L): boolean {
     return this.leftToRight.has(left);
